backend/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
beautifulsoup4==4.14.3
billiard==4.2.4
black==26.1.0
bs4==0.0.2
cbor2==5.8.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.5
daphne==4.2.1
Django==5.2.11
django-celery-beat==2.8.1
django-configurations==2.5.1
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-prometheus==2.4.1
django-stubs==5.2.9
django-stubs-ext==5.2.9
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.29.0
feedparser==6.0.12
filelock==3.20.3
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pathspec==1.0.4
pgvector==0.4.2
pillow==12.1.1
pip-review==1.3.0
platformdirs==4.9.2
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pytest==9.0.2
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.1
PyYAML==6.0.3
redis==7.1.1
referencing==0.37.0
regex==2026.1.15
rich==14.3.2
rpds-py==0.30.0
ruff==0.15.4
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
sentry-sdk==2.54.0
service-identity==24.2.0
setuptools==82.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.1.0
Twisted==25.5.0
txaio==25.12.2
typer==0.23.0
typer-slim==0.23.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
zope.interface==8.2
